services/bot/requirements.txt
aiohttp==3.9.5
teams-ai==1.3.0
python-dotenv==1.0.1
botbuilder-core>=4.7.0
botbuilder-integration-aiohttp>=4.16.1
# azure-monitor-opentelemetry==1.6.0
# botbuilder-integration-applicationinsights-aiohttp>=4.16.1
# botbuilder-applicationinsights
